requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
Pillow
opencv-python
pydicom
optuna
plotly
kaleido
